Update spring project

+++ SpringBoot/3_ExchangeCalculator/src/main/java/ExchangeRate/dto/ResponseExchangeRate.java
@@ -1,29 +1,38 @@
+/**
+ * Response Exchange Rate
+ */
+
 package ExchangeRate.dto;
+
+
+/**
+ * @author Ritik
+ */
 
 import lombok.Data;
 
 @Data
 public class ResponseExchangeRate {
 	
-	private float exchangeRate;
+	private double exchangeRate;
 	
 	public ResponseExchangeRate() {}
 	
-	public ResponseExchangeRate(float exchangeRate) {
+	public ResponseExchangeRate(double exchangeRate) {
 		this.exchangeRate = exchangeRate;
 	}
 
 	/**
 	 * @return the exchangeRate
 	 */
-	public float getExchangeRate() {
+	public double getExchangeRate() {
 		return exchangeRate;
 	}
 
 	/**
 	 * @param exchangeRate the exchangeRate to set
 	 */
-	public void setExchangeRate(float exchangeRate) {
+	public void setExchangeRate(double exchangeRate) {
 		this.exchangeRate = exchangeRate;
 	}
 }
